route refactor

+++ routes/web.php
@@ -1,10 +1,11 @@
 <?php
 
 use App\Http\Controllers\ArticleController;
-use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
-use Illuminate\Http\Request;
 use App\Http\Controllers\HomeController;
+use App\Http\Controllers\LocaleController;
+use App\Http\Controllers\LoginController;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,21 +18,29 @@ use App\Http\Controllers\HomeController;
 |
 */
 
-Route::get('/{locale?}', [HomeController::class, 'index'])->name("home.index");
 
 Route::get('/', static function () {
-    return redirect('/en');
+    return redirect('/articles/' . app()->getLocale());
 });
+
+Route::post('/change-locale', [LocaleController::class, 'changeLocale'])->name('change.locale');
 
 Route::prefix('/articles')->group(static function () {
     Route::get('/{locale?}', [ArticleController::class, 'index'])->name('articles.index');
     Route::get('/view/{id}/{locale?}', [ArticleController::class, 'view'])->name('articles.view');
-    Route::get('/edit/{articleId}/{locale?}', [ArticleController::class, 'edit'])->name('articles.edit');
-    Route::get('/list/{locale?}', [ArticleController::class, 'listArticles'])->name('articles.list');
-    Route::get('/create/{locale?}', [ArticleController::class, 'create'])->name('articles.create');
-    Route::put('/update/{id}', [ArticleController::class, 'update'])->name('articles.update');
-    Route::delete('/delete/{id}', [ArticleController::class, 'delete'])->name('articles.delete');
-    Route::post('/submit', [ArticleController::class, 'submit'])->name('articles.submit');
+    Route::get('/edit/{articleId}/{locale?}', [ArticleController::class, 'edit'])->name('articles.edit')->middleware('auth');
+    Route::get('/list/{locale?}', [ArticleController::class, 'listArticles'])->name('articles.list')->middleware('auth');
+    Route::get('/create/{locale?}', [ArticleController::class, 'create'])->name('articles.create')->middleware('auth');
+    Route::put('/update/{id}', [ArticleController::class, 'update'])->name('articles.update')->middleware('auth');
+    Route::delete('/delete/{id}', [ArticleController::class, 'delete'])->name('articles.delete')->middleware('auth');
+    Route::post('/submit', [ArticleController::class, 'submit'])->name('articles.submit')->middleware('auth');
     Route::redirect('/view', '/articles/all');
     Route::redirect('', '/articles/all/{locale}');
-});
+})->where(['locale' => '[a-zA-Z]{2}'])->middleware('setlocale');
+
+Route::get('/login/{locale?}', [App\Http\Controllers\Auth\LoginController::class, 'index'])->name('login.index');
+Route::post('/login/submit', [App\Http\Controllers\Auth\LoginController::class, 'submit'])->name('login.submit');
+Route::post('/logout', [App\Http\Controllers\Auth\LoginController::class, 'logout'])->name('logout');
+
+
+
